+++ test/topic.test.js
@@ -1,5 +1,5 @@
 const assert = require('assert');
-
+const expect = require('chai').expect;
 const Topic = require('../build/topic');
 
 // *********************************************************************
@@ -9,26 +9,30 @@ describe("Test Topic Object", function () {
   it("Test .getName()", function () {
 
     var topic = new Topic('topic-name');
-    assert.strictEqual(topic.getName('topic-name'), 'topic-name');
+    var topicName = topic.getName('topic-name');
+    expect(topicName).to.equal('topic-name');
 
   });
 
-  it("Test .getEvent()", function () {
+  it("Test .getEventList()", function () {
 
     var topic = new Topic('topic-1');
     topic.addEvent('event-1');
 
-    var result = topic.getEvent();
-    assert.strictEqual(result[0], 'event-1');
+    var eventList = topic.getEventList();
+    var eventName = eventList[0];
+    expect(eventName).to.equal('event-1');
   });
 
   it("Test .addEvent()", function () {
 
     var topic = new Topic('topic-1');
 
-    var result = topic.addEvent('my-event');
-    assert.strictEqual(result.length, 1);
-    assert.strictEqual(result[0], 'my-event');
+    var eventList = topic.addEvent('my-event');
+    expect(eventList.length).to.equal(1);
+
+    var eventName = eventList[0];
+    expect(eventName).to.equal('my-event');
   });
 
   it("Test .removeEvent()", function () {
@@ -36,8 +40,8 @@ describe("Test Topic Object", function () {
     var topic = new Topic('topic-1');
     topic.addEvent('event-1');
 
-    var result = topic.removeEvent('event-1');
-    assert.strictEqual(result.length, 0);
+    var eventList = topic.removeEvent('event-1');
+    expect(eventList.length).to.equal(0);
 
   });
 
@@ -47,10 +51,10 @@ describe("Test Topic Object", function () {
 
     topic.addEvent('some-event-name');
 
-    var result = topic.addListener('some-event-name', function listener() { });
+    var listener_id = topic.addListener('some-event-name', function listener() { });
 
     // expected result format = {event-name}/{listener-index}
-    assert.strictEqual(result, 'some-event-name/1');
+    expect(listener_id).to.equal('some-event-name/1');
 
   });
 
@@ -80,10 +84,10 @@ describe("Test Topic Object", function () {
     var topic = new Topic('topic-1');
     topic.addEvent('event-1');
 
-    var listenerId = topic.addListener('event-1', function () { });
+    var listener_id = topic.addListener('event-1', function () { });
 
-    var result = topic.removeListener(listenerId);
-    assert.strictEqual(result, true);
+    var result = topic.removeListener(listener_id);
+    expect(result).to.equal(true);
   });
 
   it("Test .dispatch()", function () {
@@ -103,9 +107,9 @@ describe("Test Topic Object", function () {
 
     topic.dispatch('event-1', 'hello word');
 
-    assert.strictEqual(dataResult.topic, 'topic-1');
-    assert.strictEqual(dataResult.event, 'event-1');
-    assert.strictEqual(dataResult.message, "hello word");
+    expect(dataResult.topic).to.equal('topic-1');
+    expect(dataResult.event).to.equal('event-1');
+    expect(dataResult.message).to.equal("hello word");
 
   });
 
@@ -120,7 +124,7 @@ describe("Test Topic Object", function () {
   });
 
   it(`Test .dispatch() : dispatch an event without a message
-      - The 'message' property of the listener argument is defined as null`, function () {
+      - The 'message' property of the listener argument is set to empty object`, function () {
 
     var dataResult = {};
 
@@ -137,9 +141,9 @@ describe("Test Topic Object", function () {
 
     topic.dispatch('event-1');
 
-    assert.strictEqual(dataResult.topic, 'topic-1');
-    assert.strictEqual(dataResult.event, 'event-1');
-    assert.strictEqual(dataResult.message, null);
+    expect(dataResult.topic).to.equal('topic-1');
+    expect(dataResult.event).to.equal('event-1');
+    expect(dataResult.message).to.be.empty;
 
   });